refactor: initial modularization
